skills/course-design-autopilot/requirements.txt
# Document parsing
mineru[full]
langdetect

# Diagram generation
plantuml

# Report generation
python-docx
docx2pdf
reportlab
